Solution to Print a Pyramid Pattern with Numbers

+++ Print a Pyramid Pattern with Numbers@@67877e1818d53eb724f435c6/code.c
@@ -1,13 +1,13 @@
 // Your code here...
 #include <stdio.h>
-double factorial(int num){
-    double fac=1;
+long long factorial(int num){
+    long long fac=1;
     for(int i=1;i<=num;i++){
         fac*=i;
     }
     return fac;
 }
-double combo(int n,int r){
+long long combo(int n,int r){
     return factorial(n)/(factorial(r)*factorial(n-r));
 }
 int main(){
